+++ mobile/android/android-components/docs/changelog.md
@@ -12,6 +12,9 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **feature-intent**
+   * ⚠️ **This is a breaking change**: `IntentProcessor.process` is not a suspend function anymore.
+
 * **feature-contextmenu**
   * Add "Share image" to context menu.